Parameterize AST, set Annotation to ast::Typing for graft module
By parameterizing the AST, we can more easily extend the syntax tree
without major refactoring of all functions that traverse it. The most
pressing annotation needed is for type inference in expressions, but we
may want positional information from the Rust AST for error messages.

Typing is delayed from the graft module for separating concerns.

The graft module only needs for the annotation to have a Default.