Fix block-resetting function to handle local declarations on top of stack
With this commit, the function `restore_stack_code` can now clean up the
stack after a code block. This will work as cleanup at the end of the
body of a while-loop, such that the stack looks identical at the start
of each while-loop iteration.

The cleanup is limited to local declarations that are on top of the
stack. The cleanup function cannot handle declarations further down, but
with the way we generate code right now, all local declarations will be
on top, so this should work for all allowable while-loop bodies.

This cleanup code should probably also be used if-expr and other code
blocks.