Generalize function to clear stack
Now this function can be used to clear the stack above a specified stack
height. It will preserve the value on top of the stack and clear
everything below it, till a specified height.

This function can be used for `return` statements and for continuing
compilation after a code block which specifies a local scope and returns
a value. So also for if/else expressions. The code does, however not
verify that the ordering of identifiers is preserved. So if the compiler
generates code inside an if/else block that changes the ordering of
previously defined variables, using this function will lead to wrong
results.

We should probably add a sanity check that the if/else expression didn't
change the ordering of bound variables on the stack.