Generalize if-expression to recover stack from assignments
This should allow the compiler to remove any local declarations that
happen within an then or else block. The code to cleanup the stack from
a then/else branch is included, and a check that previously declared
variables do not have a new position on the stack.

In `graft.rs` we don't yet allow for the compilation of these more complicated if
expressions that contain local variables but now the code generator
should be able to handle it.