Allow list-index setting through a[2] = 14u32
This is achieved by making a new data type for list indices, unknown
integer, that is then resolved by the type checker. It's not the
cleanest solution but it allows us to write code that would also be
valid Rust syntax. The tasm-lib snippet `Set` is used to set values at a
list index, and the call to that snippet is inlined in the tasm-code
generator for assignments.