Make test-helpers memory-comparison optional
Now that variable-spilling has been introduced, we want to give the
option of relaxing the test criteria, to not compare VM memory after the
execution of compiled code with some expected code. This relaxation is
introduced because memory-spilling is considered an internal
compiler-decision and should not necessarily be part of a test
framework.