Add `usize` as alias for `u32` for this compiler
In TASM, the word size is either a B-field element or a u32. When
indexing into lists, the natural data type to use is u32, so for
indexing purposes, we consider the words size to be u32. This alias is
needed to ensure that the code we are writing is a valid subset of Rust,
since the Rust compiler can only index into a list or array with a
`usize` data type, and this compiler can only index into a list with a
`u32` type. Once the index is a variable, as is the case for the test
program `usize_as_alias_for_u32` included in this commit, this compiler
cannot do type inference, so the variable needs an explicit type. The
only explicit type that works for the Rust compiler is `usize`, and the
only type that works for this compiler is `u32`. So there was no other
option than to make `usize` an alias for `u32`.