Add support for `count_ones` for `u32`
Since we have a `pop_count` instruction in TVM now, we just use that
instruction directly, we don't import a snippet from tasm-lib to achieve
this.

That change, calling a method that's not a snippet, required some
changes to the `unsigned_integers` library module. With those changes, I
think we are slowly moving towards a common interface for the libraries.