Loop over all libraries instead of listing them individually
This will make it very easy to add new functionality without having the
whole view of the compiler.

This approach does, however, create opaque behavior if multiple library
functions/methods are called the same, or if a library function is
called the same as a local function that e.g. is called recursively.