Fix evaluation order of arguments to `hash_pair`
Unfortunately `hash_pair` from Tip5 is defined such that the elements
are placed in the wrong order on the stack/sponge state if they are
evaluated left-to-right. So we add a parameter to the function
signatutre `arg_evaluation_order` that allows this function to have its
arguments evaluated in oppossite order to save instructions and clock
cycles by not having to swap the digests on the stack before calling the
`hash` instruction.