Generalize handling of method calls
Now the receiver of a method can be an expression. This was needed in
the `leaf_index_to_mt_index_and_peak_index_rast` function, which is a
helper function for the canonical MMR function
`calculate_new_peaks_from_leaf_mutation` which is the next function that
should be added to the test suite.