Divide compiler state into function state and global state
This is needed for being able to handle local function declarations
without the two functions (local and outer) overwrite each other's
memory. The `vstack` must be reset for a local function call, but the
`counter` and the `snippet_state` (which includes the static memory
allocator) must be be carried over into a local function.