Add fake support for references in source code
This is added because the function `hash_pair` takes references to
`Digest`, and references weren't supported in tasm-lang. This commit
just ignoreres the reference `&` if it sees one and grafts the inner
expression. This means that some programs will be valid tasm-lang
programs without being valid Rust programs, so we are not working with a
true subset. I'm not sure if that's a problem. I can't see why it would
be.