Add new test type inspired by `omnizk`
Adds a new way of testing a function. This test compiles and runs the function
in question on the host machine's native architecture and then compiles and
runs it in Triton VM like we've done so far. The benefit of this new testing
method is that we can ensure that we are starting from valid Rust code.
Previously we were only testing that the Rust code parsed, not the it compiled.