Let `compose` wrap compiled code in function call
Move this wrapping in a function call from test helper functions to the
code generation. This wrapping needs to be done with the current
code generations, since only through this wrapping can we call `recurse`
on the outermost layer. Otherwise, the generated code would cause a
jump stack underflow when a recurse was encountered on the outermost
level.