Allow recursice calls from methods
The method compiler needs to be made aware of its own existence. Since
methods are now located in the ftable but rather in the
`compiled_methods` field in global state, we need to add it there,
*before* we attempt to compile the method. Otherwise, the compilation
would run into an infinite loop if a method was defined recursively. To
avoid that, we add a dummy value to this field and then correct that
dummy value after the method has been compiled.