Prevent memory overwriting in OZK struct method test
Memory spilling started at address 1 and thus overwrote the struct
that was non-deterministically loaded into memory. This is fixed
by moving the object to higher addresses in RAM such that the compiler
can spill to the lower addresses, *and* by initializing the dynamic
memory allocator to the first word after where the struct is stored.