Remove overflow checks from `pow` method
The `pow` method when called on a `u32` uses the `pow` instruction in
Triton VM. This instruction has scary wrap-around behavior which you
cannot fully control for using the compiler, as the instruction is
implemented with `BFieldElement`'s `mod_pow` method which wraps around
the B-field's quotient. So since we cannot do complete overflow checks,
we elect to do zero overflow check and require the developer to know
what they're doing if they choose to use this method.