Give local functions access to functions in outer scope
When the type checker evaluates a local function, the evaluation of the
local function can now see the functions in the outer scope. This is
necessary if those local function need to call a previously declared
function.

For now, functions must be declared in the order they are used, though.