WIP: Start work on array support
Also change AST-type for index into a list/array from ast::Expr to
either ast::Expr or u32. This is done because we are only allowed to
index into an array with dynamic values *if* that array is boxed. If the
array is *not* boxed, then we can only index into it with statically
known values.