Add check for subroutines: (name1 == name2) => (code1 == code2)
If using the static allocator, it's very easy to generate two
subroutines with the same name, that do *not* share the same code. This
is the cause for a bug that occurs when you put the same enum type on
stack twice. This extra check should prevent this class of errors from
occurring again.