Change code for match-arm bindings into a subroutine
The match-arm binding code can be reused, so we include it as a
subroutine. This also prevents us from constructing incompatible
match-arm bindings if more than one of the same binding pattern occurs
throughout the program, as the code uses the static memory allocator
which can change the compiler state in an incompatible way.

This commit only covers the bindings for when the value is boxed. Next
commit will cover the same problem for stack-enums.