Improve encapsulation of custom and imported types
Add a data struct `CompositeTypes` to hold all composite types, both
those imported from libraries (like `core`) and those the programmer
has defined. This allows us to extract all type-defined constructors
that are available, such as `Foo(100)` and `Foo::Bar(200)` in a nice
encapsulated way. In the future we should probably add all methods
and associated functions to this data structure as well.