Resolve correct constructors for multiple `Result` instances
Use type hints and resulting types from the type annotation to identify
the correct enum variant constructors when there are multiple to pick
from. The test case was for multiple instances of calls to `Err(())`
which had to result in the correct `Result` type, like `Result<Bfe>` or
`Result<Xfe>`.