Fix Option support by rewriting AST in type checker
This is somewhat hacky, and I'm not sure I want to keep it. With this
patch, the AST is rewritten in the type checker such that a literal `None`
is rewritten to a enum variant declaration `Option::None`.

I think we need to handle this case in the type checker, since that's the
only place where the relevant type information is present, but I'm not sure
this handling is good.