Prevent construction of `Vec<T>` if `T` is not `Copy`
Cf. #58, `Vec<T>` can currently only be correctly constructed if the
element type implements `Copy`. Instead of fixing this and introducing a
very big runtime punishment due to re-allocations and index arithmetic
with quadratic runtimes, we simply mark the `push` method to these
vectors as `unimplemented` notifying the developer that this is
currently not supported. We don't foresee that we need these vectors for
the recufier, so we see no reason to fix it now.