Reduce all visibility to maximum `pub(crate)`
With the exception of *one* function in `lib.rs` that needs to be
visible to the `main` function, all types, traits, and functions are
made only visible to the crate through `pub(crate)`, or have even more
restricted visibility. This allows us to easily identify parts of the
code that is unused.

This change allowed me to delete a few helper functions, as the compiler
could now warn about unused code.