Introduce Tip5WithState and add SpongeState to `rust_shadows` VM state
Instead of compiling code with the Tip5 sponge state as an explicit
variable, we add the Tip5 sponge state to the monadic state used by the
Rust shadowing functions. This has the benefit that we don't need to
tell the `tasm-lang` compiler to ignore the `Tip5State` bindings present
in the source code since it's no longer there.

Thanks to Ferdinand for sending me in this direction.

Co-authred-by: Jan Ferdinand Sauer <ferdinand@neptune.cash>