refactor: Move most library-specific type grafting into libraries
Previously, we were matching on types like `Vec` and `Box` directly in
the grafter. This commit makes an abstration out of that handling and
delegates it to appropriate libraries.

Types such as `BFieldElement`, `XFieldElement`, and `Digest` are still,
kinda, considered primitive types though, so some of that type handling
still happens in the main `graft.rs` file. It could be argued that these
types should *also* be handled by libraries, but I don't consider this a
high priority, as these data types are natively supported in Triton VM.