test: With successful verification of a TVM proof!
The `recufy` function now manages to successfully verify a proof from
Triton-VM! Open the champagne!

Two caveats, though: There seems to be a bug in Triton-VM where the
randomizer polynomial is not handled correctly, and the clock cycle
count is higher than we had hoped for, at 5.3 million.

Next priroity should be to profile this execution and to make a
benchmark for it. Then we can, hopefully, start bringing that number
down!