refactor(test::recufier): Improve separation of tests and helper functions
Move all functions that only serve as entrypoints for tests under a
`#[cfg(test)]` annotation to make it easier for developers to reuse code
from the `recufier` directory. The whole `recufier` directory is
technically all test-code for this compiler, but for now we actually use
these tests to build the recufier that Triton-VM needs.